requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
flake8==7.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1